[HW13] Move server out of app
