+++ hw12_13_14_15_calendar/cmd/calendar/main.go
@@ -1,13 +1,18 @@
 package main
 
 import (
+	"context"
 	"flag"
 	"fmt"
 	"github.com/r-egorov/otus_golang/hw12_13_14_15_calendar/internal/app"
 	"github.com/r-egorov/otus_golang/hw12_13_14_15_calendar/internal/config"
 	"github.com/r-egorov/otus_golang/hw12_13_14_15_calendar/internal/logger"
+	internalhttp "github.com/r-egorov/otus_golang/hw12_13_14_15_calendar/internal/server/http"
 	"log"
 	"os"
+	"os/signal"
+	"syscall"
+	"time"
 )
 
 var configFilePath string
@@ -38,7 +43,34 @@ func main() {
 	logg := logger.New(logOut, conf.Logger.Level)
 
 	calendar := app.New(logg, conf)
-	calendar.Run()
+
+	ctx, cancel := signal.NotifyContext(context.Background(),
+		syscall.SIGINT, syscall.SIGTERM, syscall.SIGHUP)
+	defer cancel()
+
+	httpserver := internalhttp.NewServer(logg, calendar, conf.Server.Host, conf.Server.Port)
+
+	serverStopped := make(chan struct{})
+	go func() {
+		<-ctx.Done()
+
+		ctx, cancel := context.WithTimeout(context.Background(), time.Second*3)
+		defer cancel()
+
+		if err := httpserver.Stop(ctx); err != nil {
+			logg.Error("failed to stop http server: " + err.Error())
+		}
+		serverStopped <- struct{}{}
+	}()
+
+	logg.Info("calendar is running...")
+
+	if err := httpserver.Start(ctx); err != nil {
+		logg.Error("failed to start http server: " + err.Error())
+		cancel()
+		os.Exit(1)
+	}
+	<-serverStopped
 }
 
 func getLogWriter(c config.Config) (out *os.File, outClose func() error) {
